Remove code duplication between ReferenceAreaRenderer and FocusedGeometryRenderer
